+++ src/inspect_swe/_codex_cli/codex_cli.py
@@ -38,6 +38,7 @@ def codex_cli(
     system_prompt: str | None = None,
     mcp_servers: Sequence[MCPServerConfig] | None = None,
     disallowed_tools: list[Literal["web_search"]] | None = None,
+    attempts: int | AgentAttempts = 1,
     model: str | None = None,
     filter: GenerateFilter | None = None,
     cwd: str | None = None,
@@ -59,6 +60,7 @@ def codex_cli(
         system_prompt: Additional system prompt to append to default system prompt.
         mcp_servers: MCP servers to make available to the agent.
         disallowed_tools: Optionally disallow tools (currently only web_search).
+        attempts: Configure agent to make multiple attempts.
         model: Model name to use (defaults to main model for task).
         filter: Filter for intercepting bridged model requests.
         cwd: Working directory to run codex cli within.
@@ -74,8 +76,8 @@ def codex_cli(
     # resolve model
     model = f"inspect/{model}" if model is not None else "inspect"
 
-    # only support a single attempt for now as codex exec resume doesn't seem to work properly
-    attempts = AgentAttempts(attempts=1)
+    # resolve attempts
+    attempts = AgentAttempts(attempts) if isinstance(attempts, int) else attempts
 
     # ensure disallowed_tools list
     disallowed_tools = disallowed_tools or []
@@ -167,15 +169,14 @@ def codex_cli(
             debug_output: list[str] = []
             agent_prompt = prompt
             attempt_count = 0
-            resume_rollout: str | None = None
             while True:
-                # resume if requested
-                agent_cmd = cmd.copy()
-                if resume_rollout is not None:
-                    agent_cmd.extend(["-c", f'experimental_resume="{resume_rollout}"'])
-
                 # append prompt
+                agent_cmd = cmd.copy()
                 agent_cmd.append(agent_prompt)
+
+                # resume if requested
+                if attempt_count > 0:
+                    agent_cmd.extend(["resume", "--last"])
 
                 # run agent
                 result = await sbox.exec(
@@ -213,7 +214,6 @@ def codex_cli(
 
                 # otherwise update prompt with incorrect message and continue
                 else:
-                    resume_rollout = await _last_rollout(sbox, codex_home, user)
                     if callable(attempts.incorrect_message):
                         if not is_callable_coroutine(attempts.incorrect_message):
                             raise ValueError(